Delete scalePixels() method and FilterQuality in Pixmap.
