+++ DEPS
@@ -76,14 +76,9 @@
          "dir": "third_party/concurrentqueue"
       },
       {
-         "url": "https://github.com/nothings/stb.git",
-         "commit": "f58f558c120e9b32c217290b80bad1a0729fbb2c",
-         "dir": "third_party/stb"
-      },
-      {
-         "url": "https://github.com/google/highway.git",
-         "commit": "a523516d35e22a4ba8e2e70a319062cb87352de6",
-         "dir": "third_party/highway"
+        "url": "https://github.com/google/highway.git",
+        "commit": "a523516d35e22a4ba8e2e70a319062cb87352de6",
+        "dir": "third_party/highway"
       }
     ]
   },
