+++ src/platform/apple/BitmapContextUtil.h
@@ -22,7 +22,5 @@
 #include "tgfx/core/ImageInfo.h"
 
 namespace tgfx {
-uint32_t GetBitmapInfo(AlphaType alphaType, ColorType colorType);
-
 CGContextRef CreateBitmapContext(const ImageInfo& info, void* pixels);
 }  // namespace tgfx
